Remove imagej-maven-plugin from dev profile
Unfortunately, Maven does not support the "LATEST" special version for
plugins, only for dependencies. It is uncommon to need to couple to a
dev version of the imagej-maven-plugin anyway -- and if you do, you
can still do it by adding an explicit blurb to your POM:

  <properties>
    <imagej-maven-plugin.version>x.y.z-SNAPSHOT</imagej-maven-plugin.version>
  </properties>

LATEST was supported for plugins in Maven 2.x, but support was removed
in Maven 3.x. And version ranges for plugins are not supported either...