Manage imagej-scripting, not imagej-tutorials
After further consideration, it will not work to scrunch all the ImageJ
tutorial Java code into a single project -- even in src/test/java.
This is because some tutorials (e.g., the custom-preprocessor-plugin)
add rather disruptive extensions to the application context.

However, there does need to be a single component that ships all
tutorial scripts (some of which are adapted from imagej-tutorials)
in src/main/resources/script_templates/Tutorials for use with ImageJ's
Script Editor.

As an aside: for several reasons, it also does not make sense to include
Java sources as Script Editor templates. There is definitely still a
need for the imagej-tutorials multi-module project, with well-organized
examples usable from IDEs such as Eclipse.

So in a nutshell:
* Keep imagej-tutorials split up.
* Do not ship those components with ImageJ.
* Add imagej-scripting, and ship it with ImageJ.

And imagej-scripting has the added benefit of unioning all the supported
SciJava script languages into a single component, so that imagej/imagej
does not need to declare them all explicitly anymore.