Exclude com.sun:tools dep from net.imagej:ij
The tools library is a system scope dependency of ImageJ 1.x,
for compilation purposes only. It should really be declared as
"provided + system" but you cannot do that with Maven.

In any case, we do not want anything to inherit it transitively.
So the proper fix here is an exclusion on net.imagej:ij.

See also:
https://github.com/scijava/pom-scijava/commit/ff8038335f676bca29a744aab33717ca5dd5487f